requirements.txt
prometheus-client
mysql-connector-python